Created commands pkg and debug class
Created the commnds package to hold all commands and the debug class to hold debug commands specifically.
