+++ .classpath
@@ -2,9 +2,7 @@
 <classpath>
 	<classpathentry kind="src" path="src"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
-	<classpathentry kind="lib" path="E:/MC Stuff/Libraries/denizen-1.0.1-SNAPSHOT.jar"/>
-	<classpathentry kind="lib" path="E:/MC Stuff/Libraries/citizens-2.0.22-SNAPSHOT.jar"/>
-	<classpathentry kind="lib" path="E:/MC Stuff/Libraries/Vault.jar"/>
-	<classpathentry kind="lib" path="E:/MC Stuff/Libraries/Spigot/spigot-1.12.jar"/>
+	<classpathentry kind="lib" path="C:/Users/Evan/Documents/Minecraft Stuff/Plugin Workspace/Libraries/Spigot/spigot-1.12.jar"/>
+	<classpathentry kind="lib" path="C:/Users/Evan/Documents/Minecraft Stuff/Plugin Workspace/Libraries/denizen-1.0.1-SNAPSHOT.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
